requirements.txt
streamlit
openai
tiktoken
python-dotenv
requests
groq